PMM-3978 fix tests.
